reverie/requirements.txt
openai>=0.27.0
python-dotenv
pyaudio
webrtcvad
pyttsx3
